requirements.txt
python-dotenv
requests
pandas
scikit-learn